requirements.txt
streamlit
torch
torchvision
numpy
pillow
scikit-learn
matplotlib
seaborn
tqdm
